branch first HART out of others
This file is actually not a boot loader.

Basically, when machine get powered on, the bootloader (built in
hardware) will try to load OS from 0x8000_0000 location in memory.
This is where this boot.S, or its binary version, should be stored at.

For simplicity, HazelnutOS currently decided not to take advantage
of ohter HARTs. So lead other non-first HARTs goes to `wfi` and
endless loop